x64-exp git-svn-id: https://casparcg.svn.sourceforge.net/svnroot/casparcg/server/branches/64bit-exp@1846 362d55ac-95cf-4e76-9f9a-cbaa9c17b72d

+++ core/mixer/audio/audio_mixer.cpp
@@ -70,7 +70,7 @@ struct audio_mixer::implementation
 	std::stack<core::frame_transform>	transform_stack_;
 	std::map<const void*, audio_stream>	audio_streams_;
 	std::vector<audio_item>				items_;
-	std::vector<size_t>					audio_cadence_;
+	std::vector<int>					audio_cadence_;
 	video_format_desc					format_desc_;
 	
 public:
